Improve data/map loading OOB and overflow checks

+++ src/engine/shared/datafile.cpp
@@ -249,7 +249,7 @@ void *CDataFileReader::GetDataImpl(int Index, int Swap)
 {
 	if(!m_pDataFile) { return 0; }
 
-	if(Index < 0 || ((unsigned) Index) >= m_pDataFile->m_Header.m_NumRawData)
+	if(Index < 0 || Index >= m_pDataFile->m_Header.m_NumRawData)
 		return 0;
 
 	// load it if needed
@@ -315,7 +315,7 @@ void *CDataFileReader::GetDataSwapped(int Index)
 
 void CDataFileReader::ReplaceData(int Index, char *pData)
 {
-	if(Index < 0 || ((unsigned) Index) >= m_pDataFile->m_Header.m_NumRawData)
+	if(Index < 0 || Index >= m_pDataFile->m_Header.m_NumRawData)
 		return;
 
 	// make sure the data has been loaded
@@ -327,7 +327,7 @@ void CDataFileReader::ReplaceData(int Index, char *pData)
 
 void CDataFileReader::UnloadData(int Index)
 {
-	if(Index < 0 || ((unsigned) Index) >= m_pDataFile->m_Header.m_NumRawData)
+	if(Index < 0 || Index >= m_pDataFile->m_Header.m_NumRawData)
 		return;
 
 	mem_free(m_pDataFile->m_ppDataPtrs[Index]);
